hidpp10: Add default profile page in mouse info
Different mice load their default profile from different locations. The
default profile page is added to the HID++1.0 mouse info so that the
first profile can be written in the right place.

Only set the value for the G500 and G500s.
More informations are needed for the G9, G9x, G700, G700s.